pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dy-sdr-encoder"
version = "0.1.0"
description = "A package for training dynamic Sparse Distributed Representations using Hebbian learning"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "DY SDR Encoder Team"}
]
keywords = ["sdr", "sparse-distributed-representations", "hebbian-learning", "neural-networks"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.23",
    "typer>=0.9",
    "pyyaml",
    "tqdm>=4.64",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
]

[project.scripts]
dy-sdr-train = "dy_sdr_encoder.cli:app"

[project.urls]
Homepage = "https://github.com/your-org/dy-sdr-encoder"
Repository = "https://github.com/your-org/dy-sdr-encoder"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
dy_sdr_encoder = ["py.typed", "configs/*.yaml", "data/*"] 